backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
groq==0.4.2
httpx==0.27.2
pydantic==2.5.0
python-multipart==0.0.6
